Improve the lat-lon mode
In the old version, the random latitude spans from -90 to 90 degrees, which is improper because there are very few airports in the polar regions. In the new version, the random latitude will only span from the minimum latitude to the maximum latitude of all the airports selected by the checkbox.